# specs/001-backend-spec/contracts/task-api-openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: A REST API for managing user tasks with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
paths:
  /health:
    get:
      summary: Health check endpoint
      description: Verify API is running and database is connected
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  database:
                    type: string
                    example: "connected"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-01-09T10:00:00Z"
        '503':
          description: API or database is unhealthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "unhealthy"
                  database:
                    type: string
                    example: "disconnected"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-01-09T10:00:00Z"
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for a user
      description: Retrieve all tasks for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by status
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: sort
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [created, title]
            default: created
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a single task
      description: Retrieve details of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: ID of the task to retrieve
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch or task belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a task
      description: Update an existing task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch or task belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a task
      description: Permanently delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch or task belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggle the completion status of a task (completed ↔ incomplete)
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID from JWT token (must match)
          schema:
            type: string
        - name: task_id
          in: path
          required: true
          description: ID of the task to toggle
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch or task belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "user_123"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          example: "Write API specs and implementation guide"
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-01-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-09T10:00:00Z"
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "Milk, bread, eggs, coffee"
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Updated task title"
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: "Updated description"
        completed:
          type: boolean
          example: true
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "user_123"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          example: "Write API specs and implementation guide"
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-01-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-09T10:00:00Z"
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        count:
          type: integer
          example: 2
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Human-readable error message"
        code:
          type: string
          example: "ERROR_CODE_CONSTANT"